fix rocket screen fuel texture not changing always show fuel texture + fix allowing the player to remove tank and motor upgrade when the rocket is non-empty

+++ common/src/main/java/com/st0x0ef/stellaris/client/screens/OxygenDistributorScreen.java
@@ -51,7 +51,7 @@ public class OxygenDistributorScreen extends AbstractContainerScreen<OxygenDistr
             return;
         }
 
-        energyGauge.update(blockEntity.getWrappedEnergyContainer().getStoredEnergy());
+        energyGauge.update((int)blockEntity.getWrappedEnergyContainer().getStoredEnergy());
     }
 
     @Override
